refactor: extract html modules
