+++ packages/client/global.d.ts
@@ -1,4 +1,1 @@
-declare function postcss(
-  template: readonly string[] | ArrayLike<string>,
-): string;
-declare function html(template: readonly string[] | ArrayLike<string>): string;
+/// <reference types="../../scripts/types.d.ts" />
