+++ packages/client/src/setupClient.ts
@@ -1,4 +1,4 @@
-import { append, create } from './utils/dom';
+import { append, jsx } from './utils/html';
 import { defineElements } from './elements/defineElements';
 import { CLIENT, InternalElements } from './constants';
 import { Options, setOptions } from './options';
@@ -8,6 +8,6 @@ export function setupClient(options: Options) {
   if (!document.querySelector(InternalElements.HTML_INSPECT_ELEMENT)) {
     setOptions(options);
     defineElements();
-    append(document.body, create(InternalElements.HTML_INSPECT_ELEMENT));
+    append(document.body, jsx(InternalElements.HTML_INSPECT_ELEMENT));
   }
 }
