+++ scripts/types.d.ts
@@ -1,0 +1,4 @@
+declare function postcss(
+  template: readonly string[] | ArrayLike<string>,
+): string;
+declare function html(template: readonly string[] | ArrayLike<string>): string;
